Move preset-loading animation into its own file
