+++ js/repl/src/ReplOptions.js
@@ -7,6 +7,7 @@ import {
   pluginConfigs,
   presetPluginConfigs
 } from './PluginConfig';
+import PresetLoadingAnimation from './PresetLoadingAnimation';
 import Svg from './Svg';
 import { colors, media } from './styles';
 
@@ -119,7 +120,7 @@ class ExpandedContainer extends Component {
                   event.target.checked
                 )}
             />
-            {envPresetState.isLoading ? <LoadingAnimation /> : 'Enabled'}
+            {envPresetState.isLoading ? <PresetLoadingAnimation /> : 'Enabled'}
           </label>
           <div className={styles.envPresetColumn}>
             <label
@@ -255,24 +256,8 @@ const PluginToggle = ({
         toggleSetting(config.package, event.target.checked)}
       type="checkbox"
     />
-    {state.isLoading ? <LoadingAnimation /> : label || config.label}
+    {state.isLoading ? <PresetLoadingAnimation /> : label || config.label}
   </label>;
-
-const LoadingAnimation = () =>
-  <div className={styles.loadingAnimation}>
-    <div className={`${styles.loadingTick} ${styles.loadingTick1}`} />
-    <div className={`${styles.loadingTick} ${styles.loadingTick2}`} />
-    <div className={`${styles.loadingTick} ${styles.loadingTick3}`} />
-    <div className={`${styles.loadingTick} ${styles.loadingTick4}`} />
-    <div className={`${styles.loadingTick} ${styles.loadingTick5}`} />
-  </div>;
-
-const bounce = css.keyframes({
-  '0%': { transform: 'scaleY(0.25)' },
-  '40%': { transform: 'scaleY(0.75)' },
-  '80%': { transform: 'scaleY(0.25)' },
-  '100%': { transform: 'scaleY(0.25)' }
-});
 
 const styles = {
   wrapper: css({
@@ -426,38 +411,5 @@ const styles = {
     '&:disabled': {
       opacity: 0.5
     }
-  }),
-  loadingAnimation: css({
-    height: '2rem',
-    display: 'flex',
-    alignItems: 'center',
-    marginLeft: '0.5rem'
-  }),
-  loadingTick: css({
-    width: '4px',
-    height: '100%',
-    backgroundColor: 'rgba(255,255,255,0.25)',
-    display: 'inline-block',
-    animationName: bounce,
-    animationDuration: '1.4s',
-    animationIterationCount: 'infinite',
-    animationTimingFunction: 'ease-in-out',
-    marginLeft: '6px'
-  }),
-  loadingTick1: css({
-    animationDelay: 0,
-    marginLeft: 0
-  }),
-  loadingTick2: css({
-    animationDelay: '-1.1s'
-  }),
-  loadingTick3: css({
-    animationDelay: '-1.0s'
-  }),
-  loadingTick4: css({
-    animationDelay: '-0.9s'
-  }),
-  loadingTick5: css({
-    animationDelay: '-0.8s'
   })
 };
